Initial line: Add pre-request functionality in Flask server
Detailed Description:
A pre-request hook was introduced to the Flask server to execute actions before each request. With this change, a function gets implemented that prints a message and assigns the absolute path of the working directory to the global object. Also, the path was added to the homepage output, displaying it for every request.