feat: Added new error endpoint and updated relevant tests
In this update, a new error endpoint was added to the server which simply returns a status code of 500 accompanied by an HTML response. In addition, the respective test for the new error endpoint has been written ensuring that on a GET request, a status code of 500 is indeed returned. Minor modifications were also made to the order of import statements and some spaces in the redirect function's docstring.