Refactored import statements
Moved the imports related to the database connection and initialization to the beginning of the script to prevent circular import issues. This necessitated the relocation of the model import statements. Also, adjusted metadata configuration in the alembic configuration to look for models in separate files. In addition, a migration script has been added to create tables for users and products. Also, absolute imports have been clarified in model files.