feat: Transformed user_routes to return JSON instead of HTML
Overall aim of the changes made in this commit is to transform the server's user routes to return JSON instead of HTML. This includes the paths for fetching all users, a user by their id, a random user, and a selected number of random users. This will make the server's API more useful for client applications. Any error messages are also updated to be returned as JSON. The routes have also been updated to utilize a versioned path for the API.